# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  models.AddDomainRequest:
    properties:
      domain:
        type: string
    required:
    - domain
    type: object
...
  models.SystemInfo:
    properties:
      cpu_cores:
        type: integer
      cpu_name:
        type: string
      cpu_usage:
        type: string
      daily_bandwidth:
        type: string
      domain:
        type: string
      ip:
        type: string
      kernel:
        type: string
      monthly_bandwidth:
        type: string
      os:
        type: string
      ram_total_mb:
        type: integer
      ram_usage_percent:
        type: string
      ram_used_mb:
        type: integer
      uptime:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for VPN management system supporting SSH, VMESS, VLESS, Trojan,
    and Shadowsocks protocols
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: VPN API
  version: "1.0"
paths:
...
  /domain/add:
    post:
      consumes:
      - application/json
      description: Add a new domain to the system for VPN services
      parameters:
      - description: Domain name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Domain added successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Failed to add domain
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Add a new domain
      tags:
      - Domain
  /domain/current:
    get:
      description: Get the currently configured domain and server IP
      produces:
      - application/json
      responses:
        "200":
          description: Current domain retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Failed to get current domain
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current domain
      tags:
      - Domain
  /domain/ssl/renew:
    post:
      description: Renew the SSL certificate for the current domain
      produces:
      - application/json
      responses:
        "200":
          description: SSL certificate renewed successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Failed to renew SSL certificate
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Renew SSL certificate
      tags:
      - Domain
  /system/bandwidth:
    get:
      description: Get daily and monthly bandwidth usage
      produces:
      - application/json
      responses:
        "200":
          description: Bandwidth usage retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Failed to get bandwidth usage
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get bandwidth usage
      tags:
      - System
  /system/info:
    get:
      description: Get detailed system information including CPU, RAM, disk usage,
        and network info
      produces:
      - application/json
      responses:
        "200":
          description: System information retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SystemInfo'
              type: object
        "500":
          description: Failed to get system information
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get system information
      tags:
      - System
  /system/reboot:
    post:
      description: Reboot the entire system
      produces:
      - application/json
      responses:
        "200":
          description: System reboot initiated
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Failed to reboot system
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Reboot system
      tags:
      - System
  /system/restart:
    post:
      description: Restart all managed VPN services
      produces:
      - application/json
      responses:
        "200":
          description: VPN services restarted successfully
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Failed to restart services
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Restart VPN services
      tags:
      - System
  /system/status:
    get:
      description: Get the status of all managed VPN services
      produces:
      - application/json
      responses:
        "200":
          description: Service status retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Failed to get service status
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get VPN service status
      tags:
      - System
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"